[0.21] Create Cookbook class
- Added an empty Cookbook class
- Added the Classifier class with its basic functions in development
Note: As for now, the Cookbook class will be supposed to inherit from abstract classes: Classifier and FileMenager (names may differ in later versions). The first one will contain of methods for Recipes to be sorted and filtered. The second one will contain of methods converting Recipes into files and managing them in the file system.